+++ tests/test_interface/test_numpy.py
@@ -53,7 +53,19 @@ def test_numpy_coercion(model_blank):
     assert isinstance(instance.array, np.ndarray)
 
 
+@pytest.mark.scalar
+@pytest.mark.serialization
 def test_numpy_empty_string():
+    """
+    Empty strings are coerced to arrays and serialzied as such.
+
+    Mildly redundant with test_serialize_scalars_as_arrays below,
+    but a specific regression test for issue #52
+
+    References:
+        - https://github.com/p2p-ld/numpydantic/issues/52
+    """
+
     class MyModel(BaseModel):
         array: NDArray[Any, np.str_]
 
@@ -62,3 +74,27 @@ def test_numpy_empty_string():
     assert inst.array == np.array([""])
     dumped = inst.model_dump_json()
     assert json.loads(dumped)["array"] == [""]
+
+
+@pytest.mark.serialization
+@pytest.mark.scalar
+@pytest.mark.parametrize("scalar", ("", 0, 0.5))
+@pytest.mark.parametrize("as_json", (True, False))
+def test_serialize_scalars_as_arrays(scalar, as_json: bool):
+    """
+    The numpy interface matches scalar values, coerces them to arrays,
+    and serializes them as arrays in both python and json.
+    """
+
+    class MyModel(BaseModel):
+        array: NDArray
+
+    inst = MyModel(array=scalar)
+    assert isinstance(inst.array, np.ndarray)
+    assert inst.array == np.array([scalar])
+    if as_json:
+        dumped = inst.model_dump_json()
+        assert json.loads(dumped)["array"] == [scalar]
+    else:
+        dumped = inst.model_dump()
+        assert dumped["array"] == np.array([scalar])